namada_proof_of_stake
Module epoched

Epoched and EpochedDelta are structures for data that is set for future (and possibly past) epochs.

Structs

Epoched

Discrete epoched data handle

EpochedDelta

Delta epoched data handle

OffsetDefaultNumPastEpochs

Default offset

OffsetMaxProposalPeriod

Offset at max proposal period.

OffsetMaxProposalPeriodOrSlashProcessingLen

Offset at the larger of the slash processing length and the max proposal period.

OffsetMaxProposalPeriodOrSlashProcessingLenPlus

Offset at the larger of the slash processing length and the max proposal period, plus the default num past epochs.

OffsetMaxProposalPeriodPlus

Offset at the max proposal period, plus the default num past epochs.

OffsetMaxU64

Maximum offset.

OffsetPipelineLen

Offset at pipeline length.

OffsetPipelinePlusUnbondingLen

Offset at pipeline + unbonding length.

OffsetSlashProcessingLen

Offset at the slash processing delay.

OffsetSlashProcessingLenPlus

Offset at the slash processing delay plus the default num past epochs.

OffsetUnbondingLen

Offset at unbonding length.

OffsetZero

Zero offset

Enums

DynEpochOffset

Offset length dynamic choice.

Constants

DEFAULT_NUM_PAST_EPOCHS

Default number of past epochs to keep.

LAST_UPDATE_SUB_KEY

Sub-key for an epoched data structure’s last (most recent) epoch of update

LAZY_MAP_SUB_KEY

Sub-key holding a lazy map in storage

OLDEST_EPOCH_SUB_KEY

Sub-key for an epoched data structure’s oldest epoch with some data

Traits

EpochOffset

Which offset should be used to set data. The value is read from PosParams.

Type Aliases

NestedEpoched

Discrete epoched data handle with nested lazy structure