namada_proof_of_stake
Module queries

Queriezzz

Functions

bonds_and_unbonds

Collect the details of all bonds and unbonds that match the source and validator arguments.

find_bonds

Find raw bond deltas for the given source and validator address.

find_delegation_validators

Find all validators to which a given bond owner (or source) has a delegation

find_delegations

Find all validators to which a given bond owner (or source) has a delegation with the amount

find_unbonds

Find raw unbond deltas for the given source and validator address.

get_consensus_validator_from_protocol_pk

Lookup data about a validator from their protocol signing key.

get_total_voting_power

Lookup the total voting power for an epoch.

get_validator_eth_hot_key

Get a validator’s Ethereum hot key from storage at the given epoch.

get_validator_protocol_key

Find the protocol key of the given validator address at the given epoch.

has_bonds

Find if the given source address has any bonds.

read_validator_stake

Read PoS validator’s stake (sum of deltas). For non-validators and validators with 0 stake, this returns the default - token::Amount::zero().