namada_proof_of_stake
Module rewards

PoS rewards distribution.

Structs

PosRewards

Holds coefficients for the three different ways to get PoS rewards

PosRewardsCalculator

Holds relevant PoS parameters and is used to calculate the coefficients for the rewards

PosRewardsRates

Return values of the inflation asnd staking rewards rates

Enums

RewardsError

Errors during rewards calculation

Functions

add_rewards_to_counter

Add tokens to a rewards counter.

apply_inflation

Apply inflation to the Proof of Stake system.

compute_current_rewards_from_bonds

Compute the current available rewards amount due only to existing bonds. This does not include pending rewards held in the rewards counter due to unbonds and redelegations.

compute_inflation

Compute PoS inflation amount

estimate_staking_reward_rate

Compute an estimation of the most recent staking rewards rate.

read_rewards_counter

Read the current token value in the rewards counter.

take_rewards_from_counter

Take tokens from a rewards counter. Deletes the record after reading.

update_rewards_products_and_mint_inflation

Update validator and delegators rewards products and mint the inflation tokens into the PoS account.