namada_proof_of_stake
Module slashing

Slashing tingzzzz

Functions

apply_list_slashes

Computes how much remains from an amount of tokens after applying a list of slashes.

compute_amount_after_slashing_unbond

Compute the total amount of tokens from a set of unbonds, both redelegated and not, after applying slashes. Used in unbond_tokens.

compute_amount_after_slashing_withdraw

Compute the total amount of tokens from a set of unbonds, both redelegated and not, after applying slashes. Used in withdraw_tokens.

compute_bond_at_epoch

Get the remaining token amount in a bond after applying a set of slashes.

compute_slash_bond_at_epoch

Uses fn compute_bond_at_epoch to compute the token amount to slash in order to prevent overslashing.

compute_slashable_amount

Computes how much is left from a bond or unbond after applying a slash given that a set of slashes may have been previously applied.

find_all_enqueued_slashes

Collect the details of all of the enqueued slashes to be processed in future epochs into a nested map

find_all_slashes

Find all slashes and the associated validators in the PoS system

find_slashes_in_range

Find slashes applicable to a validator with inclusive start and exclusive end epoch.

find_validator_slashes

Find PoS slashes applied to a validator, if any

get_slashed_amount

Compute a token amount after slashing, given the initial amount and a set of slashes. It is assumed that the input slashes are those committed while the amount was contributing to voting power.

process_slashes

Process enqueued slashes that were discovered earlier.

slash

Record a slash for a misbehavior that has been received from Tendermint and then jail the validator, removing it from the validator set. The slash rate will be computed at a later epoch.

slash_redelegation

Slash a redelegated bond.

slash_validator

Compute slash amounts for a validator.

slash_validator_redelegation

Slash a redelegated bond on a destination validator.