namada_proof_of_stake::slashing
Function compute_amount_after_slashing_withdraw

pub fn compute_amount_after_slashing_withdraw<S>(
    storage: &S,
    params: &OwnedPosParams,
    unbonds_and_redelegated_unbonds: &BTreeMap<(Epoch, Epoch), (Amount, EagerRedelegatedBondsMap)>,
    slashes: Vec<Slash>,
) -> Result<ResultSlashing>
where
    S: StorageRead,

Compute the total amount of tokens from a set of unbonds, both redelegated and not, after applying slashes. Used in withdraw_tokens.