namada_proof_of_stake::slashing
Function slash_redelegation

pub fn slash_redelegation<S>(
    storage: &S,
    params: &OwnedPosParams,
    amount: Amount,
    bond_start: Epoch,
    redel_bond_start: Epoch,
    src_validator: &Address,
    current_epoch: Epoch,
    slashes: &Slashes,
    total_redelegated_unbonded: &TotalRedelegatedUnbonded,
    slash_rate: Dec,
    slashed_amounts: &mut BTreeMap<Epoch, Amount>,
) -> Result<()>
where
    S: StorageRead,

Slash a redelegated bond.

Computes how many tokens will be slashed from a redelegated bond, considering that the bond may have been completely or partially unbonded and that the source validator may have misbehaved within the redelegation slashing window.