namada_proof_of_stake::slashing
Function slash_validator_redelegation

pub fn slash_validator_redelegation<S>(
    storage: &S,
    params: &OwnedPosParams,
    src_validator: &Address,
    current_epoch: Epoch,
    outgoing_redelegations: &NestedMap<Epoch, LazyMap<Epoch, Amount>>,
    slashes: &Slashes,
    dest_total_redelegated_unbonded: &TotalRedelegatedUnbonded,
    slash_rate: Dec,
    dest_slashed_amounts: &mut BTreeMap<Epoch, Amount>,
) -> Result<()>
where
    S: StorageRead,

Slash a redelegated bond on a destination validator.

In the context of a redelegation, the function computes how much a validator (the destination validator of the redelegation) should be slashed due to the misbehaving of a second validator (the source validator of the redelegation). The function computes how much the validator would be slashed at all epochs between the current epoch (curEpoch) + 1 and the current epoch + 1 + PIPELINE_OFFSET, accounting for any tokens of the redelegation already unbonded.

src_validator - the source validator

outgoing_redelegations - a map from pair of epochs to int that includes all the redelegations from the source validator to the destination validator.

The outer key is epoch at which the bond started at the source validator.

The inner key is epoch at which the redelegation started (the epoch at which was issued).

slashes a list of slashes of the source validator.

dest_total_redelegated_unbonded - a map of unbonded redelegated tokens at the destination validator.

slash_rate - the rate of the slash being processed.

dest_slashed_amounts - a map from epoch to already processed slash amounts.

Adds any newly processed slash amount to dest_slashed_amounts.