namada_proof_of_stake
Module storage

PoS functions for reading and writing to storage and lazy collection handles associated with given storage_keys.

Functions

below_capacity_validator_set_handle

Get the storage handle to the epoched below-capacity validator set

bond_handle

Get the storage handle to a bond, which is dynamically updated with when unbonding

consensus_validator_set_handle

Get the storage handle to the epoched consensus validator set

delegation_targets_handle

Get the storage handle to the delegation targets map

delegator_redelegated_bonds_handle

Get the storage handle to a delegator’s redelegated bonds information

delegator_redelegated_unbonds_handle

Get the storage handle to a delegator’s redelegated unbonds information

enqueued_slashes_handle

Get the storage handle to list of all slashes to be processed and ultimately placed in the validator_slashes_handle

find_validator_by_raw_hash

Get the validator address given the raw hash of the Tendermint consensus key

get_consensus_key

Find a consensus key of a validator account.

get_consensus_key_set

Get the unique set of consensus keys in storage

get_last_reward_claim_epoch

Get the last epoch in which rewards were claimed from storage, if any

get_num_consensus_validators

Count the number of consensus validators

is_consensus_key_used

Check if the given consensus key is already being used to ensure uniqueness.

liveness_missed_votes_handle

Get the storage handle to the missed votes for liveness tracking

liveness_sum_missed_votes_handle

Get the storage handle to the sum of missed votes for liveness tracking

read_all_validator_addresses

Read all validator addresses.

read_below_capacity_validator_set_addresses

Read all addresses from below-capacity validator set.

read_below_capacity_validator_set_addresses_with_stake

Read all addresses from below-capacity validator set with their stake.

read_below_threshold_validator_set_addresses

Read all addresses from the below-threshold set

read_consensus_validator_set_addresses

Read all addresses from consensus validator set.

read_consensus_validator_set_addresses_with_stake

Read all addresses from consensus validator set with their stake.

read_last_block_proposer_address

Read last block proposer address.

read_last_pos_inflation_amount

Read last epoch’s PoS inflation amount.

read_last_staked_ratio

Read last epoch’s staked ratio.

read_non_pos_owned_params

Read non-PoS-owned parameters to add them to OwnedPosParams to construct PosParams.

read_owned_pos_params

Read owned PoS parameters

read_pos_params

Read PoS parameters

read_total_active_stake

Read PoS total stake (sum of deltas).

read_total_stake

Read PoS total stake (sum of deltas).

read_validator_avatar

Read PoS validator’s avatar.

read_validator_deltas_value

Read PoS validator’s delta value.

read_validator_description

Read PoS validator’s description.

read_validator_discord_handle

Read PoS validator’s discord handle.

read_validator_email

Read PoS validator’s email.

read_validator_last_slash_epoch

Read the most recent slash epoch for the given epoch

read_validator_max_commission_rate_change

Read PoS validator’s max commission rate change.

read_validator_metadata

Read validator’s metadata.

read_validator_name

Read PoS validator’s name.

read_validator_stake

Read PoS validator’s stake (sum of deltas). For non-validators and validators with 0 stake, this returns the default - token::Amount::zero().

read_validator_state

Read the validator state

read_validator_website

Read PoS validator’s website.

rewards_accumulator_handle

Get the storage handle to the rewards accumulator for the consensus validators in a given epoch

total_active_deltas_handle

Get the storage handle to the total active deltas

total_bonded_handle

Get the storage handle to a validator’s total bonds, which are not updated due to unbonding

total_consensus_stake_handle

Get the storage handle to the total consensus validator stake

total_deltas_handle

Get the storage handle to the total deltas

total_unbonded_handle

Get the storage handle to a validator’s total-unbonded map

try_insert_consensus_key

Check if the given consensus key is already being used to ensure uniqueness.

unbond_handle

Get the storage handle to an unbond

update_total_deltas

Update PoS total deltas. Note: for EpochedDelta, write the value to change storage by

update_validator_deltas

Add or remove PoS validator’s stake delta value

validator_addresses_handle

Get the storage handle to the set of all validators

validator_commission_rate_handle

Get the storage handle to a PoS validator’s commission rate

validator_consensus_key_handle

Get the storage handle to a PoS validator’s consensus key (used for signing block votes).

validator_deltas_handle

Get the storage handle to a PoS validator’s deltas

validator_eth_cold_key_handle

Get the storage handle to a PoS validator’s eth cold key.

validator_eth_hot_key_handle

Get the storage handle to a PoS validator’s eth hot key.

validator_incoming_redelegations_handle

Get the storage handle to a validator’s incoming redelegations

validator_outgoing_redelegations_handle

Get the storage handle to a validator’s outgoing redelegations

validator_protocol_key_handle

Get the storage handle to a PoS validator’s protocol key key.

validator_rewards_products_handle

Get the storage handle to a validator’s rewards products

validator_set_positions_handle

Get the storage handle to a PoS validator’s deltas

validator_slashes_handle

Get the storage handle to a PoS validator’s slashes

validator_state_handle

Get the storage handle to a PoS validator’s state

validator_total_redelegated_bonded_handle

Get the storage handle to a validator’s total redelegated bonds

validator_total_redelegated_unbonded_handle

Get the storage handle to a validator’s outgoing redelegations

write_last_block_proposer_address

Write last block proposer address.

write_last_pos_inflation_amount

Write last epoch’s pos inflation amount.

write_last_reward_claim_epoch

Write the last epoch in which rewards were claimed for the delegator-validator pair

write_last_staked_ratio

Write last epoch’s staked ratio.

write_pos_params

Write PoS parameters

write_validator_address_raw_hash

Write PoS validator’s address raw hash.

write_validator_avatar

Write PoS validator’s avatar. If the provided arg is an empty string, remove the data.

write_validator_description

Write PoS validator’s description. If the provided arg is an empty string, remove the data.

write_validator_discord_handle

Write PoS validator’s discord handle. If the provided arg is an empty string, remove the data.

write_validator_email

Write PoS validator’s email. The email cannot be removed, so an empty string will result in an error.

write_validator_last_slash_epoch

Write the most recent slash epoch for the given epoch

write_validator_max_commission_rate_change

Write PoS validator’s max commission rate change.

write_validator_metadata

Write validator’s metadata.

write_validator_name

Write PoS validator’s name. If the provided arg is an empty string, remove the data.

write_validator_website

Write PoS validator’s website. If the provided arg is an empty string, remove the data.