namada_proof_of_stake
Module storage_key

Proof-of-Stake storage keys and storage integration.

Constants

VALIDATOR_STORAGE_PREFIX

Functions

below_capacity_validator_set_key

Storage key for below-capacity validator set

bond_key

Storage key for a bond with the given ID (source and validator).

bonds_for_source_prefix

Storage key prefix for all bonds of the given source address.

bonds_prefix

Storage key prefix for all bonds.

consensus_keys_key

Storage key for consensus keys set.

consensus_validator_rewards_accumulator_key

Storage key for the consensus validator set rewards accumulator.

consensus_validator_set_key

Storage key for consensus validator set

delegation_targets_key

Storage key for the delegation targets of a delegator.

delegation_targets_prefix

Storage prefix for the delegation targets.

delegator_redelegated_bonds_key

Storage key for a particular delegator’s redelegated bond information.

delegator_redelegated_bonds_prefix

Storage key prefix for all delegators’ redelegated bonds.

delegator_redelegated_unbonds_key

Storage key for a particular delegator’s redelegated unbond information.

delegator_redelegated_unbonds_prefix

Storage key prefix for all delegators’ redelegated unbonds.

enqueued_slashes_key

Storage key for all slashes.

get_validator_address_from_bond

Get validator address from bond key

is_below_capacity_validator_set_key

Is storage key for the below-capacity validator set?

is_bond_epoched_meta_key

Is storage key for a bond last update or oldest epoch? Returns the bond ID if so.

is_bond_key

Is storage key for a bond? Returns the bond ID and bond start epoch if so.

is_consensus_keys_key

Is storage key for consensus keys set?

is_consensus_validator_set_accumulator_key

Is storage key for the consensus validator set?

is_consensus_validator_set_key

Is storage key for the consensus validator set?

is_delegation_targets_key

Is storage key for the delegation targets of a delegator?

is_delegator_redelegations_key

Is the storage key’s prefix matching delegator’s total redelegated bonded or unbond amounts? If so, returns the delegator’s address.

is_last_block_proposer_key

Is storage key for block proposer address of the previous block?

is_last_pos_reward_claim_epoch_key

Is the storage key for epoch at which an account last claimed PoS inflationary rewards? Return the bond ID if so.

is_params_key

Is storage key for PoS parameters?

is_pos_key

Is the given key a PoS storage key?

is_rewards_counter_key

Is the storage key for rewards counter?

is_total_active_deltas_key

Is storage key for total active deltas?

is_total_consensus_stake_key

Is storage key for the total consensus stake?

is_total_deltas_key

Is storage key for total deltas of all validators?

is_unbond_key

Is storage key for an unbond? Returns the bond ID and unbond start and withdraw epoch if it is.

is_validator_address_raw_hash_key

Is storage key for validator’s address raw hash?

is_validator_addresses_key

Is the storage key a prefix for all active validators?

is_validator_commission_rate_key

Is storage key for validator’s commission rate?

is_validator_consensus_key_key

Is storage key for validator’s consensus key?

is_validator_deltas_key

Is storage key for validator’s total deltas?

is_validator_eth_cold_key_key

Is storage key for validator’s eth cold key?

is_validator_eth_hot_key_key

Is storage key for validator’s eth hot key?

is_validator_last_known_product_epoch_key

Is storage key for validator’s last known rewards product epoch?

is_validator_max_commission_rate_change_key

Is storage key for validator’s maximum commission rate change per epoch?

is_validator_metadata_key

Is storage key for some piece of validator metadata?

is_validator_redelegations_key

Is the storage key’s prefix matching one of validator’s:

is_validator_rewards_product_key

Is storage key for validator’s rewards products?

is_validator_set_positions_key

Is the storage key for validator set positions?

is_validator_slashes_key

Is storage key for a validator’s slashes

is_validator_state_epoched_meta_key

Is storage key for a validator state’s last update or oldest epoch?

is_validator_state_key

Is storage key for validator’s state?

is_validator_total_bond_or_unbond_key

Is the storage key for the total bonds or unbonds for a validator?

last_block_proposer_key

Storage key for block proposer address of the previous block.

last_pos_inflation_amount_key

Storage key for the last epoch’s PoS inflation amount.

last_pos_reward_claim_epoch_key

Storage key for epoch at which an account last claimed PoS inflationary rewards.

last_pos_reward_claim_epoch_prefix

Storage prefix for epoch at which an account last claimed PoS inflationary rewards.

last_staked_ratio_key

Storage key for the last epoch’s staked ratio.

liveness_data_prefix

Storage prefix for the liveness data of the cosnensus validator set.

liveness_missed_votes_key

Storage key for the liveness records.

liveness_sum_missed_votes_key

Storage key for the liveness data.

params_key

Storage key for PoS parameters.

rewards_counter_key

Storage key for rewards counter.

rewards_counter_prefix

Storage prefix for rewards counter.

slashes_prefix

Storage prefix for slashes.

total_active_deltas_key

Storage key for total active deltas (Consensus, Below-Capacity, and Below-threshold validators).

total_consensus_stake_key

Storage key for total consensus stake

total_deltas_key

Storage key for total deltas of all validators.

unbond_key

Storage key for an unbond with the given ID (source and validator).

unbonds_for_source_prefix

Storage key prefix for all unbonds of the given source address.

unbonds_prefix

Storage key prefix for all unbonds.

validator_address_raw_hash_key

Storage key for validator’s address raw hash for look-up from raw hash of an address to address.

validator_addresses_key

Storage prefix for all active validators (consensus, below-capacity, below-threshold, inactive, jailed)

validator_avatar_key

Storage key for a validator’s avatar

validator_commission_rate_key

Storage key for validator’s commission rate.

validator_consensus_key_key

Storage key for validator’s consensus key.

validator_deltas_key

Storage key for validator’s deltas.

validator_description_key

Storage key for a validator’s description

validator_discord_key

Storage key for a validator’s discord handle

validator_email_key

Storage key for a validator’s email

validator_eth_cold_key_key

Storage key for validator’s eth cold key.

validator_eth_hot_key_key

Storage key for validator’s eth hot key.

validator_incoming_redelegations_key

Storage key for a validator’s incoming redelegations, where the prefixed validator is the destination validator.

validator_last_known_product_epoch_key

Storage key for validator’s last known rewards product epoch.

validator_last_slash_key

Storage key for the last (most recent) epoch in which a slashable offense was detected for a given validator

validator_max_commission_rate_change_key

Storage key for validator’s maximum commission rate change per epoch.

validator_name_key

Storage key for a validator’s name

validator_outgoing_redelegations_key

Storage key for a validator’s outgoing redelegations, where the prefixed validator is the source validator.

validator_rewards_product_key

Storage key for validator’s rewards products.

validator_set_positions_key

Storage key for validator set positions

validator_sets_prefix

Storage prefix for validator sets.

validator_slashes_key

Storage key for validator’s slashes.

validator_state_key

Storage key for validator’s consensus key.

validator_total_bonded_key

Storage key for the total bonds for a given validator.

validator_total_redelegated_bonded_key

Storage key for validator’s total-redelegated-bonded amount to track for slashing

validator_total_redelegated_unbonded_key

Storage key for validator’s total-redelegated-unbonded amount to track for slashing

validator_total_unbonded_key

Storage key for validator’s total-unbonded amount to track for slashing

validator_website_key

Storage key for a validator’s website