namada_proof_of_stake
Trait LazyCollection

pub trait LazyCollection {
    type SubKey: Debug;
    type Value: BorshDeserialize + Debug;

    // Required methods
    fn open(key: Key) -> Self;
    fn is_valid_sub_key(&self, key: &Key) -> Result<Option<Self::SubKey>, Error>;
    fn is_data_sub_key(&self, key: &Key) -> bool;
}

A lazy collection of storage values.

This is a handler with some storage prefix that is given to its fn new(). The values are typically nested under this prefix and they can be changed individually (e.g. without reading in the whole collection) and their changes directly indicated to the validity predicates, which do not need to iterate the whole collection pre/post to find diffs.

An empty collection must be deleted from storage.

Required Associated Types

type SubKey: Debug

Possible sub-keys in the collection

type Value: BorshDeserialize + Debug

A type of a value in the inner-most collection

Required Methods

fn open(key: Key) -> Self

Create or use an existing vector with the given storage key.

fn is_valid_sub_key(&self, key: &Key) -> Result<Option<Self::SubKey>, Error>

Check if the given storage key is a valid LazyVec sub-key and if so return which one. Returns:

Ok(Some(_)) if it’s a valid sub-key

Ok(None) if it’s not a sub-key

Err(_) if it’s an invalid sub-key

fn is_data_sub_key(&self, key: &Key) -> bool

Check if the given storage key is a valid data key.

For most collections, this is the same as is_valid_sub_key, but for example for LazyVec, which has an additional sub-key for length of the vec, only the element data sub-keys would return true.

Object Safety

This trait is not object safe.