namada_proof_of_stake
Trait Read

pub trait Read<S> {
    // Required methods
    fn is_validator(storage: &S, address: &Address) -> Result<bool, Error>;
    fn is_delegator(
        storage: &S,
        address: &Address,
        epoch: Option<Epoch>,
    ) -> Result<bool, Error>;
    fn pipeline_len(storage: &S) -> Result<u64, Error>;
    fn unbonding_len(storage: &S) -> Result<u64, Error>;
    fn total_active_stake<Gov>(
        storage: &S,
        epoch: Epoch,
    ) -> Result<Amount, Error>
       where Gov: Read<S>;
    fn is_active_validator<Gov>(
        storage: &S,
        validator: &Address,
        epoch: Epoch,
    ) -> Result<bool, Error>
       where Gov: Read<S>;
    fn read_validator_stake<Gov>(
        storage: &S,
        validator: &Address,
        epoch: Epoch,
    ) -> Result<Amount, Error>
       where Gov: Read<S>;
    fn bond_amount<Gov>(
        storage: &S,
        validator: &Address,
        delegator: &Address,
        epoch: Epoch,
    ) -> Result<Amount, Error>
       where Gov: Read<S>;
}

Abstract PoS storage read interface

Required Methods

fn is_validator(storage: &S, address: &Address) -> Result<bool, Error>

Check if the provided address is a validator address

fn is_delegator( storage: &S, address: &Address, epoch: Option<Epoch>, ) -> Result<bool, Error>

Check if the provided address is a delegator address, optionally at a particular epoch. Returns false if the address is a validator.

fn pipeline_len(storage: &S) -> Result<u64, Error>

Read PoS pipeline length parameter

fn unbonding_len(storage: &S) -> Result<u64, Error>

Read PoS unbonding length parameter

fn total_active_stake<Gov>(storage: &S, epoch: Epoch) -> Result<Amount, Error>
where Gov: Read<S>,

Read total active stake

fn is_active_validator<Gov>( storage: &S, validator: &Address, epoch: Epoch, ) -> Result<bool, Error>
where Gov: Read<S>,

Returns Ok(true) if the given address is a validator and it’s not jailed or inactive

fn read_validator_stake<Gov>( storage: &S, validator: &Address, epoch: Epoch, ) -> Result<Amount, Error>
where Gov: Read<S>,

Read PoS validator’s stake. For non-validators and validators with 0 stake, this returns the default - token::Amount::zero().

fn bond_amount<Gov>( storage: &S, validator: &Address, delegator: &Address, epoch: Epoch, ) -> Result<Amount, Error>
where Gov: Read<S>,

Get the total bond amount, including slashes, for a given bond ID and epoch. Returns the bond amount after slashing. For future epochs, the value is subject to change.

Object Safety

This trait is not object safe.