namada_proof_of_stake
Trait ResultExt

pub trait ResultExt<T> {
    // Required methods
    fn into_storage_result(self) -> Result<T, Error>;
    fn wrap_err(self, msg: &'static str) -> Result<T, Error>;
}

Result extension to easily wrap custom errors into Error.

Required Methods

fn into_storage_result(self) -> Result<T, Error>

Convert a std::result::Result into storage_api Result.

fn wrap_err(self, msg: &'static str) -> Result<T, Error>

Add a static message to a possible error in Result.