namada_proof_of_stake
Module types

Proof of Stake data types

Structs

BondDetails

Bond with all its details

BondId

ID of a bond and/or an unbond.

BondsAndUnbondsDetail

Bonds and unbonds with all details (slashes and rewards, if any)

CommissionPair

Commission rate and max commission rate change per epoch for a validator

ConsensusValidator

Newly updated consensus validator’s consensus key and bonded stake.

DelegationEpochs

Contains information on epoch periods (start, end) in which a delegator had a bonded with a certain validator. The end epoch is the first epoch at which the bond ceased to exist (exclusive).

GenesisValidator

A genesis validator definition.

LivenessInfo

Liveness data related to the network and set of consensus validators

Position

A position in a validator set

Redelegation

Eager data for a generic redelegation

ResultSlashing

Temp: In quint this is from ResultUnbondTx field resultSlashing: {sum: int, epochMap: Epoch -> int}

ReverseOrdTokenAmount

A wrapper over token::Amount, whose KeySeg implementation has reverse order of the token::Amount type.

Slash

A slash applied to validator, to punish byzantine behavior by removing their staked tokens at and before the epoch of the slash.

SlashedAmount

Slashed amount of tokens.

UnbondDetails

Unbond with all its details

ValidatorLiveness

Some liveness data for a consensus validator

ValidatorMetaData

Validator metadata

VoteInfo

VoteInfo inspired from tendermint for validators whose signature was included in the last block

WeightedValidator

Validator’s address with its voting power.

Enums

SlashType

A type of slashable event.

ValidatorSetUpdate

An update of the consensus and below-capacity validator set.

ValidatorState

Validator’s state. May correspond to the validator set within which the validator belongs.

Functions

into_tm_voting_power

Calculate voting power in the tendermint context (which is stored as i64) from the number of tokens

Type Aliases

BelowCapacityValidatorSet

Below-capacity validator set, keyed by staked token amount

BelowCapacityValidatorSets

Epoched below-capacity validator sets.

Bonds

Epoched validator’s bonds

BondsAndUnbondsDetails

Bonds and unbonds with all details (slashes and rewards, if any) grouped by their bond IDs.

CommissionRates

Epoched validator commission rate

ConsensusKeys

Consensus keys set, used to ensure uniqueness

ConsensusValidatorSet

Consensus validator set, keyed by staked token amount

ConsensusValidatorSets

Epoched consensus validator sets.

DelegationTargets

The set of all target validators for a given delegator.

DelegatorRedelegatedBonded

A delegator’s redelegated bonded token amount. The map keys from outside in are:

DelegatorRedelegatedUnbonded

A delegator’s redelegated unbonded token amounts. The map keys from outside in are:

EagerRedelegatedBondsMap

In-memory map of redelegated bonds. The map keys from outside in are:

EpochedSlashes

Epoched slashes, where the outer epoch key is the epoch in which the slash is processed.

IncomingRedelegations

A validator’s incoming redelegations, where the key is the bond owner address and the value is the redelegation end epoch

LivenessMissedVotes

Liveness record of the consensus validator set. Records the block heights at which the consensus validator missed a vote.

LivenessSumMissedVotes

The sum of missed votes within some interval for each of the consensus validators.

OutgoingRedelegations

A validator’s outgoing redelegations, where the validator in question is a source validator.

RedelegatedBondsOrUnbonds

Map of redelegated bonds or unbonds. The map keys from outside in are:

RedelegatedTokens

Map of redelegated tokens. The map keys from outside in are:

RewardsAccumulator

Consensus validator rewards accumulator (for tracking the fractional block rewards owed over the course of an epoch)

RewardsProducts

Epoched rewards products

Slashes

Slashes applied to validator, to punish byzantine behavior by removing their staked tokens at and before the epoch of the slash.

TotalConsensusStakes

Epoched total consensus validator set stake

TotalDeltas

Epoched total deltas.

TotalRedelegatedBonded

A validator’s total redelegated unbonded tokens for any delegator. The map keys from outside in are:

TotalRedelegatedUnbonded

A validator’s total redelegated unbonded tokens for any delegator. The map keys from outside in are:

Unbonds

Epoched validator’s unbonds

ValidatorAddresses

An epoched lazy set of all known validator addresses

ValidatorConsensusKeys

Epoched validator’s consensus key.

ValidatorDeltas

Epoched validator’s deltas.

ValidatorEthColdKeys

Epoched validator’s eth cold key.

ValidatorEthHotKeys

Epoched validator’s eth hot key.

ValidatorPositionAddresses

A map from a position to an address in a Validator Set

ValidatorProtocolKeys

Epoched validator’s protocol key.

ValidatorSetPositions

Stored positions of validators in validator sets

ValidatorSlashes

Slashes indexed by validator address and then block height (for easier retrieval and iteration when processing)

ValidatorStateInfo

The validator state from a query and the epoch when it was queried.

ValidatorStates

Epoched validator’s state.

ValidatorTotalUnbonded

Total unbonded for validators needed for slashing computations.