namada_proof_of_stake
Module validator_set_update

Validator set updates

Functions

copy_validator_sets_and_positions

Copy the consensus and below-capacity validator sets and positions into a future epoch. Also copies the epoched set of all known validators in the network.

insert_validator_into_validator_set

Insert the new validator into the right validator set (depending on its stake)

promote_next_below_capacity_validator_to_consensus

Promote the next below-capacity validator to the consensus validator set.

remove_below_capacity_validator

Remove a validator from the below-capacity set

remove_consensus_validator

Remove a validator from the consensus validator set

update_validator_set

Update validator set at the pipeline epoch when a validator receives a new bond and when its bond is unbonded (self-bond or delegation).

validator_set_update_comet

Communicate imminent validator set updates to Tendermint.